bot/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-telegram-bot==21.5
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
